feat(gwasUnifiedFlowRefactorRebased): Created StateManagement directory, files and updated refs to them
