+++ src/Analysis/GWASV2/Shared/Covariates/CustomDichotomousCovariates.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import _ from 'lodash';
 import SelectStudyPopulation from '../../SelectStudyPopulation/SelectStudyPopulation';
-import { ACTIONS } from '../constants';
+import ACTIONS from '../StateManagement/Actions';
 import '../../../GWASUIApp/GWASUIApp.css';
 
 const CustomDichotomousCovariates = ({
