+++ src/Analysis/GWASV2/Shared/StateManagement/Actions.js
@@ -1,0 +1,14 @@
+const ACTIONS = {
+  SET_SELECTED_STUDY_POPULATION_COHORT: 'setSelectedStudyPopulationCohort',
+  DECREMENT_CURRENT_STEP: 'decrementCurrentStep',
+  INCREMENT_CURRENT_STEP: 'incrementCurrentStep',
+  SET_OUTCOME: 'setOutcome',
+  ADD_COVARIATE: 'addCovariate',
+  DELETE_CONTINUOUS_COVARIATE: 'deleteContinouosCovariate',
+  DELETE_DICHOTOMOUS_COVARIATE: 'deleteDichotomousCovariate',
+  UPDATE_MAF_THRESHOLD: 'updateMafThreshold',
+  UPDATE_IMPUTATION_SCORE: 'updateImputationScore',
+  UPDATE_NUM_PCS: 'updateNumPCs',
+};
+
+export default ACTIONS;
