+++ src/Analysis/GWASV2/Shared/StateManagement/InitialState.js
@@ -1,0 +1,13 @@
+const initialState = {
+  outcome: {},
+  selectedStudyPopulationCohort: null,
+  covariates: [],
+  imputationScore: 0.3,
+  mafThreshold: 0.01,
+  numOfPC: 3,
+  gwasName: '',
+  selectedHare: { concept_value: '' },
+  currentStep: 0,
+};
+
+export default initialState;
